diary: Daily entry
Updated diary with intentions of working on project plan and researching further. Also this commit marks the addition of a new branch plan where I'll be shaping up this repo in the early stages of the project. Also added reading_material.txt for personal reference to keep track of research material.