poc: Notebook 9 - CombinedPreprocessor
Altered K-Folds C-V Scores function to allow a preprocessor to be passed as a parameter, allowing preprocessing in between folds. Created a CombinedPreprocessor class that combines ordinal, numerical and categorical preprocessing pipelines and concatenates the transformations into one usable set. Added a label encoder to the Classification Tree that eased predictions with various kinds of labels. Implemented NumericalConverter into the preprocessing module