ui: Cross-Validation and Confusion Matrices
Added further functionality to the UI by allowing loaded datasets to be evaluated in 5-folds cross-validation, the accuracy scores are printed and then the confusion matrices from accumalated results of all folds are displayed below. Preprocessing is very limited so this only works with numeric datasets for now, such as ionosphere and german credit rating dataset. Preprocessing still needs to be automated. Added further options for user to toggle header removal and select the target column and delimiter of the loaded dataset