poc: Grid Search + Notebook 10
Added an optimisers module to the backend that allows tuning of hyperparameters to be automated for all three models. I've also managed to display all the scores during tuning in a heatmap. I've tried to demonstrate progress on this in a Notebook 10. Will have to work more on preprocessing in this case and embed preprocessing into the interface