poc: Notebook 10 - Titanic Grid Search
Augmented the GridSearch class such that it now has preset parameter grids for the three models - which means the grid doesn't have be to be initialised every single time. The grid is general enough to be appropriate for most cases but can be adjusted if need be when initialised. Displayed heatmaps and utilised best models in k-folds cross-validation