ui: Process Data Input Boxes
Removed the provious pop-up dialog boxes and replaced with input boxes on the interface itself. Added a few datasets to the repo. Added a preview of the first data sample in the dataset with the user-specified data type columns of that sample - to show or help verify if columns have been inputted correctly